Integrate dependency dump into the main script
